sdk-sample-scripts/ShodanVulns/requirements.txt
requests>=2.29.0
flatten_json>=0.1.13
runzero-sdk>=0.8.1